Separate recipe route logic into utility functions
